refactor(mm-next): move footer into shared folder
