+++ packages/mirror-media-next/pages/index.js
@@ -24,7 +24,7 @@ import { transformRawDataToArticleInfo } from '../utils'
 import EditorChoice from '../components/editor-choice'
 import LatestNews from '../components/latest-news'
 import ShareHeader from '../components/shared/share-header'
-import Footer from '../components/footer'
+import Footer from '../components/shared/footer'
 
 /**
  * @typedef {import('../components/shared/share-header').HeaderData['flashNewsData']} FlashNewsData
