+++ packages/mirror-media-next/pages/premiumsection/[slug].js
@@ -8,7 +8,7 @@ import SectionArticles from '../../components/shared/section-articles'
 import { GCP_PROJECT_ID } from '../../config/index.mjs'
 import { fetchHeaderDataInPremiumPageLayout } from '../../utils/api'
 import ShareHeader from '../../components/shared/share-header'
-import Footer from '../../components/footer'
+import Footer from '../../components/shared/footer'
 
 /**
  * @typedef {import('../../type/theme').Theme} Theme
