+++ packages/mirror-media-next/pages/tag/[slug].js
@@ -8,7 +8,7 @@ import TagArticles from '../../components/tag/tag-articles'
 import { GCP_PROJECT_ID } from '../../config/index.mjs'
 import { fetchHeaderDataInDefaultPageLayout } from '../../utils/api'
 import ShareHeader from '../../components/shared/share-header'
-import Footer from '../../components/footer'
+import Footer from '../../components/shared/footer'
 
 const TagContainer = styled.main`
   width: 320px;
