+++ packages/mirror-media-next/pages/video/[id].js
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import axios from 'axios'
 
 import ShareHeader from '../../components/shared/share-header'
-import Footer from '../../components/footer'
+import Footer from '../../components/shared/footer'
 import { fetchHeaderDataInDefaultPageLayout } from '../../utils/api'
 import { GCP_PROJECT_ID, URL_RESTFUL_SERVER } from '../../config/index.mjs'
 import {
