+++ packages/mirror-media-next/pages/video_category/[slug].js
@@ -6,7 +6,7 @@ import CategoryVideos from '../../components/video_category/category-videos.js'
 import { VIDEOHUB_CATEGORIES_PLAYLIST_MAPPING } from '../../constants/index.js'
 import styled from 'styled-components'
 import ShareHeader from '../../components/shared/share-header.js'
-import Footer from '../../components/footer'
+import Footer from '../../components/shared/footer'
 import { fetchHeaderDataInDefaultPageLayout } from '../../utils/api/index.js'
 import client from '../../apollo/apollo-client.js'
 import { fetchCategory } from '../../apollo/query/categroies.js'
